requirements.txt
awesome-streamlit==20200728.1
requests==2.25.1
thop==0.0.31.post2005241907
numpy==1.20.1
opencv_python==4.5.1.48
streamlit==0.76.0
seaborn==0.11.1
pandas==1.2.1
scipy==1.6.0
torch==1.7.1+cu110
pycocotools==2.0.2
matplotlib==3.3.4
tqdm==4.56.0
torchvision==0.8.2+cu110
coremltools==4.1
onnx==1.8.1
Pillow==8.1.0
PyYAML==5.4.1
wandb==0.10.18
